Deleted the isoSPI sheet and connections
On the schematic, I deleted the isoSPI sheet and its connections to the STM (replaced them with no connection flags). I updated the PCB from that so it now has all the components that need to be arranged and it can be done from scratch. I also added in a ground plane.